Variables Practicing

+++ 03 - Programming Using C++ - Level 1/Variables Practicing/Variables Practicing.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
     cout << "Integer Variables:\n";
-    int MyAge;
+    short MyAge;
     MyAge = 50;
     cout << MyAge << endl;
     MyAge = 23;
@@ -13,7 +13,7 @@ int main()
 
     cout << "Other Variables:\n";
 
-    int MyNumber = 50;
+    short MyNumber = 50;
     float MyFloatNumber = 7.84;
     double MyDoubleNumber = 21.66659;
     char MyLetter = 'O';
